refactor: 继续重构并细分总览界面小组件卡片
