+++ src/renderer/components/states/expand/components/OverviewTab/utils/countdown.ts
@@ -1,0 +1,6 @@
+export function cdDiffDays(targetStr: string): number {
+  const today = new Date();
+  today.setHours(0, 0, 0, 0);
+  const target = new Date(`${targetStr}T00:00:00`);
+  return Math.ceil((target.getTime() - today.getTime()) / (1000 * 60 * 60 * 24));
+}
